conda-requirements.txt
oldest-supported-numpy
torch
h5py
matplotlib
